+++ fastly/ngwaf/v1/lists/doc.go
@@ -1,0 +1,3 @@
+// Package lists contains API operations to create, get, update, list, and
+// delete Fastly Next-Gen WAF lists.
+package lists